[SERVICE-CONTROLLERS] Move MvcConfigNormalizer to Pimcore\Controller\Config\ConfigNormalizer
