+++ pimcore/lib/Pimcore/Bundle/CoreBundle/Resources/config/services.yml
@@ -1,8 +1,4 @@
 services:
-    pimcore.mvc_config_normalizer:
-        class: Pimcore\Controller\MvcConfigNormalizer
-        arguments: ['@kernel']
-
     pimcore.bundle_locator:
         class: Pimcore\HttpKernel\BundleLocator\BundleLocator
         arguments: ['@kernel']
@@ -28,12 +24,12 @@ services:
             - '@pimcore.document_service'
             - '@pimcore.service.request.site_resolver'
             - '@pimcore.http.request_helper'
-            - '@pimcore.mvc_config_normalizer'
+            - '@pimcore.controller.config.config_normalizer'
 
     # dedicated router for static routes
     pimcore.routing.router.staticroute:
         class: Pimcore\Routing\Staticroute\Router
-        arguments: ['@router.request_context', '@pimcore.mvc_config_normalizer']
+        arguments: ['@router.request_context', '@pimcore.controller.config.config_normalizer']
         calls:
             - [setLogger, ['@logger']]
         tags:
@@ -54,6 +50,14 @@ services:
             - [setLogger, ['@logger']]
         tags:
             - { name: monolog.logger, channel: routing }
+
+    #
+    # CONTROLLERS
+    #
+
+    pimcore.controller.config.config_normalizer:
+        class: Pimcore\Controller\Config\ConfigNormalizer
+        arguments: ['@kernel']
 
     #
     # SESSION
@@ -184,7 +188,7 @@ services:
 
     pimcore.service.request.template_resolver:
         class: Pimcore\Service\Request\TemplateResolver
-        arguments: ['@request_stack', '@pimcore.mvc_config_normalizer']
+        arguments: ['@request_stack', '@pimcore.controller.config.config_normalizer']
 
     pimcore.service.request.template_vars_resolver:
         class: Pimcore\Service\Request\TemplateVarsResolver
