+++ pimcore/lib/Pimcore/Routing/Dynamic/DocumentRouteHandler.php
@@ -18,7 +18,7 @@ declare(strict_types=1);
 namespace Pimcore\Routing\Dynamic;
 
 use Pimcore\Config;
-use Pimcore\Controller\MvcConfigNormalizer;
+use Pimcore\Controller\Config\ConfigNormalizer;
 use Pimcore\Http\RequestHelper;
 use Pimcore\Model\Document;
 use Pimcore\Routing\DocumentRoute;
@@ -44,7 +44,7 @@ class DocumentRouteHandler implements DynamicRouteHandlerInterface
     private $requestHelper;
 
     /**
-     * @var MvcConfigNormalizer
+     * @var ConfigNormalizer
      */
     private $configNormalizer;
 
@@ -57,13 +57,13 @@ class DocumentRouteHandler implements DynamicRouteHandlerInterface
      * @param Document\Service $documentService
      * @param SiteResolver $siteResolver
      * @param RequestHelper $requestHelper
-     * @param MvcConfigNormalizer $configNormalizer
+     * @param ConfigNormalizer $configNormalizer
      */
     public function __construct(
         Document\Service $documentService,
         SiteResolver $siteResolver,
         RequestHelper $requestHelper,
-        MvcConfigNormalizer $configNormalizer
+        ConfigNormalizer $configNormalizer
     ) {
         $this->documentService  = $documentService;
         $this->siteResolver     = $siteResolver;
